requirements.txt
joblib==1.3.2
langchain==0.0.340
numpy==1.26.2
pandas==2.1.3
scikit_learn==1.3.2
termcolor==2.3.0
torch==2.0.1
torchmetrics==1.2.0
tqdm==4.66.1
transformers==4.34.1
google-api-core==2.14.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-documentai==2.20.2
googleapis-common-protos==1.61.0
llama_cpp_python==0.2.15